server/requirements.txt
fastapi
uvicorn
textblob
pydantic
python-dotenv
email-validator
openai
scikit-learn
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
